[#818] Move configuration properties to API projects
