+++ core/api/src/main/java/com/blazebit/persistence/ConfigurationProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.blazebit.persistence.impl;
+package com.blazebit.persistence;
 
 /**
  *
