+++ core/impl/src/main/java/com/blazebit/persistence/impl/MutableQueryConfiguration.java
@@ -16,6 +16,8 @@
 
 package com.blazebit.persistence.impl;
 
+import com.blazebit.persistence.ConfigurationProperties;
+
 import java.util.Map;
 
 /**
