+++ entity-view/api/src/main/java/com/blazebit/persistence/view/ConfigurationProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.blazebit.persistence.view.impl;
+package com.blazebit.persistence.view;
 
 /**
  *
