+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/update/subview/multiparent/jointable/EntityViewUpdateSubviewMultiParentJoinTableTest.java
@@ -23,7 +23,7 @@ import com.blazebit.persistence.testsuite.entity.Document;
 import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.view.FlushMode;
 import com.blazebit.persistence.view.FlushStrategy;
-import com.blazebit.persistence.view.impl.ConfigurationProperties;
+import com.blazebit.persistence.view.ConfigurationProperties;
 import com.blazebit.persistence.view.spi.EntityViewConfiguration;
 import com.blazebit.persistence.view.testsuite.update.AbstractEntityViewUpdateDocumentTest;
 import com.blazebit.persistence.view.testsuite.update.subview.multiparent.jointable.model.DocumentIdView;
